Rename package duke to be package workdone
