+++ src/main/java/workdone/command/HelpCommand.java
@@ -1,7 +1,7 @@
-package duke.command;
+package workdone.command;
 
-import duke.data.Storage;
-import duke.data.TaskList;
+import workdone.data.Storage;
+import workdone.data.TaskList;
 
 /**
  * Represents a command that asks for help. A subclass of the Command class.
