+++ src/main/java/workdone/task/Event.java
@@ -1,4 +1,4 @@
-package duke.task;
+package workdone.task;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
