+++ src/main/java/workdone/task/Task.java
@@ -1,4 +1,4 @@
-package duke.task;
+package workdone.task;
 
 import java.time.format.DateTimeFormatter;
