+++ src/main/java/workdone/ui/Parser.java
@@ -1,29 +1,29 @@
-package duke.ui;
+package workdone.ui;
 
 import java.time.format.DateTimeParseException;
 import java.util.Stack;
 
-import duke.command.AddTaskCommand;
-import duke.command.Command;
-import duke.command.DeleteTaskCommand;
-import duke.command.ExitCommand;
-import duke.command.FindTaskCommand;
-import duke.command.GetListCommand;
-import duke.command.HelpCommand;
-import duke.command.TaskDoneCommand;
-import duke.command.TaskUndoneCommand;
-import duke.command.UndoCommand;
-import duke.exception.DukeException;
-import duke.exception.InvalidCommandException;
-import duke.exception.InvalidTaskNoException;
-import duke.exception.InvalidTimeException;
-import duke.exception.MissingCommandDetailException;
-import duke.exception.MultipleTimeSlotsException;
-import duke.exception.UnableToUndoException;
-import duke.task.Deadline;
-import duke.task.Event;
-import duke.task.Task;
-import duke.task.ToDo;
+import workdone.command.AddTaskCommand;
+import workdone.command.Command;
+import workdone.command.DeleteTaskCommand;
+import workdone.command.ExitCommand;
+import workdone.command.FindTaskCommand;
+import workdone.command.GetListCommand;
+import workdone.command.HelpCommand;
+import workdone.command.TaskDoneCommand;
+import workdone.command.TaskUndoneCommand;
+import workdone.command.UndoCommand;
+import workdone.exception.DukeException;
+import workdone.exception.InvalidCommandException;
+import workdone.exception.InvalidTaskNoException;
+import workdone.exception.InvalidTimeException;
+import workdone.exception.MissingCommandDetailException;
+import workdone.exception.MultipleTimeSlotsException;
+import workdone.exception.UnableToUndoException;
+import workdone.task.Deadline;
+import workdone.task.Event;
+import workdone.task.Task;
+import workdone.task.ToDo;
 
 /**
  * Represents a parser that deals with making sense of the user command.
