+++ src/main/resources/view/MainWindow.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.VBox?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="duke.ui.MainWindow">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="800.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="workdone.ui.MainWindow">
     <children>
         <TextField fx:id="userInput" layoutX="20.0" layoutY="560.0" onAction="#handleUserInput" prefHeight="30.0" prefWidth="650.0" />
         <Button fx:id="sendButton" layoutX="690.0" layoutY="560.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="30.0" prefWidth="80.0" text="Send" />
